+++ components/keyscreen/api/src/main/java/com/flipperdevices/keyscreen/api/emulate/ForbiddenFrequencyException.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.keyscreen.api.emulate
+
+class ForbiddenFrequencyException : RuntimeException()