+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/composable/ComposableSimpleEmulateButton.kt
@@ -14,8 +14,8 @@ import com.flipperdevices.keyscreen.api.Picture
 import com.flipperdevices.keyscreen.emulate.R
 import com.flipperdevices.keyscreen.emulate.composable.common.ComposableActionDisable
 import com.flipperdevices.keyscreen.emulate.composable.common.ComposableActionLoading
-import com.flipperdevices.keyscreen.emulate.composable.common.ComposableAlreadyOpenedAppDialog
 import com.flipperdevices.keyscreen.emulate.composable.common.ComposableEmulateButtonWithText
+import com.flipperdevices.keyscreen.emulate.composable.common.ComposableErrorDialogs
 import com.flipperdevices.keyscreen.emulate.model.DisableButtonReason
 import com.flipperdevices.keyscreen.emulate.model.EmulateButtonState
 import com.flipperdevices.keyscreen.emulate.viewmodel.SimpleEmulateViewModel
@@ -48,9 +48,7 @@ fun ComposableSimpleEmulateButton(modifier: Modifier = Modifier, flipperKey: Fli
         return
     }
 
-    if (emulateButtonState == EmulateButtonState.AppAlreadyOpenDialog) {
-        ComposableAlreadyOpenedAppDialog(emulateViewModel::closeDialog)
-    }
+    ComposableErrorDialogs(emulateButtonState, emulateViewModel::closeDialog)
 
     when (emulateButtonState) {
         is EmulateButtonState.Disabled -> ComposableActionDisable(